src/requirements.txt
azure-functions
pandas
numpy
azure-identity
azure-storage-blob